Trying out different feature extraction techniques for better results in Logistic Regression.
