+++ models/logistic/LogRegCNN/results.txt
@@ -1,0 +1,2 @@
+2025-03-31 04:02:04,294 - INFO - Script started.
+2025-03-31 04:02:26,981 - INFO - Script started.
