+++ models/logistic/LogRegHOG/LogRegHOG.py
@@ -1,32 +1,36 @@
 import os
+import cv2
 import datetime
 import numpy as np
 import pandas as pd
 import logging
 import matplotlib.pyplot as plt
 from tqdm import tqdm
+from skimage.feature import hog
 from sklearn.preprocessing import StandardScaler
 from sklearn.linear_model import LogisticRegression
 from sklearn.metrics import accuracy_score, classification_report
 from sklearn.decomposition import PCA
-from tensorflow.keras.applications import VGG16
-from tensorflow.keras.applications.vgg16 import preprocess_input
-from tensorflow.keras.preprocessing.image import ImageDataGenerator
-from tensorflow.keras.models import Model
-import zipfile
 
-# ========================== STEP 1: SETUP FILE PATHS ==========================
 script_dir = os.path.dirname(os.path.abspath(__file__))
-data_path = os.path.join(script_dir, "../../data/recognition/")
+data_path = os.path.join(script_dir, "../../../data/recognition/")
 train_csv = os.path.join(data_path, "train.csv")
 test_csv = os.path.join(data_path, "test.csv")
 logger = os.path.join(script_dir, "results.txt")
 
-# ========================== STEP 2: LANGUAGE SELECTION ==========================
 Lmap = {
-    1: "assamese", 2: "bengali", 3: "english", 4: "gujarati", 5: "hindi",
-    6: "kannada", 7: "malayalam", 8: "marathi", 9: "punjabi", 
-    10: "tamil", 11: "telugu", 12: "urdu"
+    1: "assamese",
+    2: "bengali",
+    3: "english",
+    4: "gujarati",
+    5: "hindi",
+    6: "kannada",
+    7: "malayalam",
+    8: "marathi",
+    9: "punjabi",
+    10: "tamil",
+    11: "telugu",
+    12: "urdu"
 }
 
 def select_lang():
@@ -39,64 +43,53 @@ def select_lang():
 logging.basicConfig(filename=logger, level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s", filemode="a")
 logging.info("Script started.")
 
-# ========================== STEP 3: EXTRACT DATASET IF NOT EXIST ==========================
-zip_path = os.path.join(script_dir, "../../recognition.zip")
-if not os.path.exists(data_path):  # Check if already extracted
-    print("Extracting recognition.zip...")
-    with zipfile.ZipFile(zip_path, "r") as zip_ref:
-        zip_ref.extractall(os.path.join(script_dir, "../../"))
-    print("Extraction completed!")
-
-# ========================== STEP 4: LOAD PRETRAINED CNN MODEL ==========================
-base_model = VGG16(weights='imagenet', include_top=False, input_shape=(224, 224, 3))
-model = Model(inputs=base_model.input, outputs=base_model.output)
-
-# ========================== STEP 5: BATCH PROCESSING FUNCTION ==========================
-def extract_cnn_features_batch(csv_path, selected_lang, batch_size=32):
+def load_dataset(csv_path, selected_lang):
     df = pd.read_csv(csv_path, skiprows=1, header=None, names=["image_path", "annotation", "script"])
+    x, y = [], []
+    logging.info(f"Loading dataset from {csv_path} for language: {selected_lang}")
     
-    # Fix image paths
-    df["image_path"] = df["image_path"].apply(lambda x: os.path.join(data_path, x.replace("\\", "/")))
-    
-    # Filter data for the selected language (Binary Classification)
-    df["label"] = df["image_path"].apply(lambda x: 1 if os.path.basename(os.path.dirname(x)).lower() == selected_lang else 0)
+    for _, row in tqdm(df.iterrows(), total=len(df), desc="Loading Images"):
+        img_path = os.path.join(data_path, row["image_path"].replace("\\", "/"))  
+        if not os.path.exists(img_path):
+            logging.warning(f"File not found: {img_path}")
+            continue
 
-    # Data Generator for Batch Processing
-    datagen = ImageDataGenerator(preprocessing_function=preprocess_input)
-    generator = datagen.flow_from_dataframe(
-        dataframe=df,
-        x_col="image_path",
-        y_col="label",
-        target_size=(224, 224),
-        batch_size=batch_size,
-        class_mode="raw",  # Since it's binary classification
-        shuffle=False
-    )
+        folder_name = os.path.basename(os.path.dirname(img_path)).lower()
+        label = 1 if folder_name == selected_lang else 0
 
-    # Extract CNN features in batches
-    features = model.predict(generator, verbose=1)
-    features = features.reshape(features.shape[0], -1)  # Flatten
+        img = cv2.imread(img_path, cv2.IMREAD_GRAYSCALE)
+        if img is None:
+            logging.warning(f"Error reading image: {img_path}")
+            continue
 
-    return features, np.array(df["label"])
+        img = cv2.resize(img, (64, 64))
+        hog_features = hog(img, pixels_per_cell=(8, 8), cells_per_block=(2, 2), feature_vector=True)
+        x.append(hog_features)
+        y.append(label)
 
-# ========================== STEP 6: SELECT LANGUAGE AND LOAD DATA ==========================
+    x, y = np.array(x), np.array(y)
+    unique_classes = np.unique(y)
+    if len(unique_classes) < 2:
+        logging.error(f"Only one class found: {unique_classes}. Check dataset.")
+        print(f"Error: Only one class found ({unique_classes}). Adjust dataset!")
+        exit()
+
+    return x, y
+
 selected_lang = select_lang()
 if not selected_lang:
     print("Invalid choice. Exiting program.")
     exit()
 
-# Extract features in batches
-x_train, y_train = extract_cnn_features_batch(train_csv, selected_lang)
-x_test, y_test = extract_cnn_features_batch(test_csv, selected_lang)
+x_train, y_train = load_dataset(train_csv, selected_lang)
+x_test, y_test = load_dataset(test_csv, selected_lang)
 
 logging.info(f"Train set size: {len(x_train)}, Test set size: {len(x_test)}")
 
-# ========================== STEP 7: PREPROCESS DATA ==========================
 scaler = StandardScaler()
 x_train = scaler.fit_transform(x_train)
 x_test = scaler.transform(x_test)
 
-# ========================== STEP 8: TRAIN LOGISTIC REGRESSION ==========================
 model = LogisticRegression(solver='liblinear', class_weight='balanced')
 model.fit(x_train, y_train)
 
@@ -110,7 +103,6 @@ print("Classification Report:\n", classification_report(y_test, y_pred))
 
 logging.info("Script finished.")
 
-# ========================== STEP 9: SAVE DECISION BOUNDARY PLOT ==========================
 plots_dir = os.path.join(script_dir, "plots")
 os.makedirs(plots_dir, exist_ok=True)
 
@@ -137,8 +129,9 @@ def save_plot(x, y, model, language):
     filename = f"{language}_{timestamp}.png"
     plot_path = os.path.join(plots_dir, filename)
     plt.savefig(plot_path)
-
+    
     logging.info(f"Saved decision boundary plot for {language} at {plot_path}")
+    logging.info("=========================================")   
     print(f"Plot saved at: {plot_path}")
     plt.show()
