+++ models/logistic/LogRegHOG/results.txt
@@ -197,8 +197,5 @@ weighted avg       0.97      0.86      0.91     22336
 2025-03-28 17:06:53,879 - INFO - Script finished.
 2025-03-28 17:07:00,872 - INFO - Saved decision boundary plot for telugu at d:\MyStuff\College\.4thSemester\PRML\Project\Script-Identifier\models\logistic\plots\telugu_20250328_170658.png
 2025-03-28 17:07:00,872 - INFO - =========================================
-2025-03-31 03:38:29,574 - INFO - Script started.
-2025-03-31 03:38:36,765 - INFO - Loading dataset from d:\MyStuff\College\.4thSemester\PRML\Project\Script-Identifier\models\logistic\../../data/recognition/train.csv for language: bengali
-2025-03-31 03:42:46,264 - INFO - Script started.
-2025-03-31 03:42:59,190 - WARNING - 5 out of the last 5 calls to <function TensorFlowTrainer.make_predict_function.<locals>.one_step_on_data_distributed at 0x000002CCDB5AB760> triggered tf.function retracing. Tracing is expensive and the excessive number of tracings could be due to (1) creating @tf.function repeatedly in a loop, (2) passing tensors with different shapes, (3) passing Python objects instead of tensors. For (1), please define your @tf.function outside of the loop. For (2), @tf.function has reduce_retracing=True option that can avoid unnecessary retracing. For (3), please refer to https://www.tensorflow.org/guide/function#controlling_retracing and https://www.tensorflow.org/api_docs/python/tf/function for  more details.
-2025-03-31 03:42:59,261 - WARNING - 6 out of the last 6 calls to <function TensorFlowTrainer.make_predict_function.<locals>.one_step_on_data_distributed at 0x000002CCDB5AB760> triggered tf.function retracing. Tracing is expensive and the excessive number of tracings could be due to (1) creating @tf.function repeatedly in a loop, (2) passing tensors with different shapes, (3) passing Python objects instead of tensors. For (1), please define your @tf.function outside of the loop. For (2), @tf.function has reduce_retracing=True option that can avoid unnecessary retracing. For (3), please refer to https://www.tensorflow.org/guide/function#controlling_retracing and https://www.tensorflow.org/api_docs/python/tf/function for  more details.
+2025-03-31 04:26:09,482 - INFO - Script started.
+2025-03-31 04:27:16,234 - INFO - Script started.
